@@ -1,26 +1,22 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import "./App.css";
+import { BatchUpdate } from "./components/batch-update";
+import { SuspensePage } from "./components/suspense";
+import { StartTransitionPage } from "./components/start-transition";
+import { DeferredValuePage } from "./components/deferred-value";
 
-function App() {
+// createRoot
+// Batch update - flushSync
+// Suspense
+// startTransition -- useDeferredValue
+// Streaming ssr
+
+export const App = () => {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div>
+      {/* <BatchUpdate /> */}
+      {/* <SuspensePage /> */}
+      {/* <StartTransitionPage /> */}
+      <DeferredValuePage />
     </div>
   );
-}
-
-export default App;
+};
